+++ linux-2.6.37.4/include/curse/curse.h
@@ -95,15 +95,15 @@ struct syscurse {
 
 //Function prototypes (although forwards are ugly:)). : All the functions return 0 for success, or one of the usual error codes for error.
 int syscurse_list_all(void);
-int syscurse_activate(int);
-int syscurse_deactivate(int);
-int syscurse_check_curse_activity(int);
+int syscurse_activate(uint64_t);
+int syscurse_deactivate(uint64_t);
+int syscurse_check_curse_activity(uint64_t);
 int syscurse_check_tainted_process(pid_t);
-int syscurse_cast(int, pid_t);
-int syscurse_lift(int, pid_t);
+int syscurse_cast(uint64_t, pid_t);
+int syscurse_lift(uint64_t, pid_t);
 int syscurse_show_rules(void);
-int syscurse_add_rule(int, char *);
-int syscurse_rem_rule(int, char *);
+int syscurse_add_rule(uint64_t, char *);
+int syscurse_rem_rule(uint64_t, char *);
 
 #include "curse_sched.h"	//Source it here too.
